src/requirements.txt
requests
dnspython
shodan
geocoder
tabulate
python-whois